refactor of auth, tabs, and header.
Changed the way accounts were being verified as an admin, restructured the header to be part of appcomponent and turned the tabs into a fully fleshed component with its own module. This was all necessary in order to fix issues with hiding elements, refreshing elements automatically when necessary, and just making life easier labeling items for specific roles.

refactored out a lot of code that I initially liked, but in use did not feel necessary, such as loading when updating an account.

added padding to classes so they would not appear under the header, could not find a better solution.

hid elements in matches and players that should be hidden when a user is not admin.

routing changes as necessary.

updated the guard to hopefully prevent access if not admin, has not been fully implemented.

removal of old header and tabs.

moved some services to a common location.

added auth service to work in coordination with token service.