fixes and finalization of auth implementation
login, register, and account view are all now in modals instead of their own page.

If the user tries to force loading a webpage they shouldn't be able to, they will be redirected to home. Implemented the user role directive.

added a module that can be imported, which exports user role directive, so that it is easily available in multiple modules.

Fixes:
Users now refreshes when a user is added or edited.
when a user logs out and another logs in, the profile will display the correct data.
added a global css for modals so that they will not take the full screen.
removed duplicate route for users, players, and home
changed html for tabs to solve routing issue
fixed a problem with duplicate tabs after removing duplicate routes
small edit to guards, still not using canLoad. Will likely go unused.